Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-	printf("int: %lu \n", sizeof(int));
-	printf("float: %lu \n", sizeof(float));
-	printf("char: %lu \n", sizeof(char));
-	printf("double: %lu \n", sizeof(double));
+	printf("int: %lu \n", (unsigned long) sizeof(int));
+	printf("float: %lu \n",(unsigned long) sizeof(float));
+	printf("char: %lu \n",(unsigned long) sizeof(char));
+	printf("double: %lu \n",(unsigned long) sizeof(double));
 	return (0);
 }
